ipaplatform: allow to configure rndc
ISC BIND may be supplied with rndc configuration.
By default named(9.16):
> If no controls statement is present, named sets up a default control
channel listening on the loopback address 127.0.0.1 and its IPv6
counterpart, ::1. In this case, and also when the controls statement is
present but does not have a keys clause, named attempts to load the
command channel key from the file rndc.key in /etc (or whatever
sysconfdir was specified when BIND was built). To create an rndc.key
file, run rndc-confgen -a.

For example, ALTLinux provides a custom configuration path, while
Fedora uses the defaults.