alt: nsswitch.conf: make sss the primary service for automount
RHEL-based distros rely on default master map configuration which
includes option '+auto.master':
> Additionally, a map may be included from its source as if it were
  itself present in the master map by including a line of the form:
  +[maptype,format:]map[options]
Thereby, all the services specified in nsswitch.conf are searched for
master map whilst only a first found master map will be used if
'+auto.master' is not set.

ALTLinux's default configuration for autofs doesn't contain
'+auto.master'. This and the incorrect order of nsswitch configuration
for automount results in IPA's automount mappings are ignored.

With this change `ipa-client-automount` sets sss as the primary
service for automount on installation.