ALT: tox: install local IPA packages
There is no priority of package source for pip. Hence,
it's undefined which package is installed, it can be resolved as
either local built wheel or PyPI's one. This happens if package
has same name and version. Upstream makes use of git snapshots
for CI builds while in ALTLinux we use release versions and thus,
if upstream published the same version as we have in downstream
pip installs PyPI's IPA and CI fails on tox task because upstream
doesn't have support for ALTLinux yet.

Fix:
don't query for IPA's packages names and install wheels instead.