Align startup_timeout with the systemd default and document it
We had it set to 120 seconds while the systemd default is 90.
They should be the same because the first one that times out "wins".

Move where during the installation we create the systemd override
file so that the timeout will be applied across all subsequent
server starts during and post installation.

Fixes: https://pagure.io/freeipa/issue/9743

Signed-off-by: Rob Crittenden <rcritten@redhat.com>
Reviewed-By: Florence Blanc-Renaud <flo@redhat.com>
Reviewed-By: Alexander Bokovoy <abokovoy@redhat.com>