Add token options to immutables for pki override
Setting the HSM options in a pki-override-file will
result in a successful installation but it affects the
storage of the token library path. It is stored as None.

We could also instead save the provided value from the
override but an override isn't necessary in this case
since IPA provides command-line options to enable it.

Signed-off-by: Rob Crittenden <rcritten@redhat.com>
Reviewed-By: Florence Blanc-Renaud <flo@redhat.com>